+++ tests/acceptance_tests/test_backtest_cached.py
@@ -47,13 +47,13 @@ class TestBacktestAcceptanceTestsWithCache:
         self.usdjpy = TestInstrumentProvider.default_fx_ccy("USD/JPY")
 
         self.engine.add_instrument(self.usdjpy)
-        self.engine.add_bars(
+        self.engine.add_bars_as_ticks(
             self.usdjpy.id,
             BarAggregation.MINUTE,
             PriceType.BID,
             TestDataProvider.usdjpy_1min_bid()[:2000],
         )
-        self.engine.add_bars(
+        self.engine.add_bars_as_ticks(
             self.usdjpy.id,
             BarAggregation.MINUTE,
             PriceType.ASK,